+++ website/app/docs/_meta.js
@@ -1,0 +1,8 @@
+export default {
+  index: "Overview",
+  "getting-started": "Getting started",
+  "core-concepts": "Core concepts",
+  api: "API reference",
+  security: "Security",
+  utils: "Micro-UI Utils",
+};